Member and Maintance are quite done

+++ P2P Tool Library/database/migrations/2026_05_10_000000_initial_database_sync.php
@@ -13,7 +13,7 @@ return new class extends Migration
             Schema::create('categories', function (Blueprint $table) {
                 $table->integer('id', true);
                 $table->string('name', 70);
-                $table->integer('parent_id');
+                $table->unsignedBigInteger('parent_id');
             });
         }
 
@@ -205,3 +205,10 @@ return new class extends Migration
         // Keep as is for safety
     }
 };
+
+
+
+
+
+
+
